# server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Todo API
  version: 1.0.0
  description: Complete Todo application with CRUD operations
servers:
  - url: http://localhost:3000

paths:
  /api/tasks:
    get:
      summary: Get all tasks
      tags: [Tasks]
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    
    post:
      summary: Create a new task
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTask'
      responses:
        '201':
          description: Task created successfully
        '400':
          description: Title is required

  /api/tasks/{id}:
    put:
      summary: Update a task
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTask'
      responses:
        '200':
          description: Task updated
    
    delete:
      summary: Delete a task
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    
    CreateTask:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
    
    UpdateTask:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean